requirements_ct.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.2.0
blinker==1.8.2
blosc2==2.0.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
cmake==3.31.1
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
databricks-sdk==0.38.0
Deprecated==1.2.15
docker==7.1.0
exceptiongroup==1.2.2
filelock==3.13.1
flake8==7.1.1
Flask==3.0.3
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==4.13.0
importlib_resources==6.4.5
iniconfig==2.0.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lit==18.1.8
Mako==1.3.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
portalocker==3.0.0
propcache==0.2.0
protobuf==5.29.1
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyDeprecate==0.3.1
pyflakes==3.2.0
pyparsing==3.1.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==1.4.5
pytz==2024.2
PyYAML==6.0.2
ray==2.10.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
sympy==1.13.3
tables==3.8.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.0.1
torch-ema==0.3
torchmetrics==0.6.0
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.0
yacs==0.1.8
yarl==1.15.2
zipp==3.20.2
